requirements.txt
matplotlib==3.8.4
numpy==1.26.4
opencv-python-headless==4.9.0.80
RPi.GPIO==0.7.1
tensorflow==2.16.1
Flask==3.0.3